requirements.txt
PyYAML==6.0
robotframework==4.1.3
attrs==21.4.0
importlib-metadata==4.10.1
more-itertools==8.12.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
six==1.16.0
wcwidth==0.2.5
zipp==3.7.0
